ai-osint-tool/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-whois==0.8.0
fake-useragent==1.4.0
